requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
sumy==0.11.0
nltk==3.8.1
Werkzeug==3.0.1
Flask-CORS==4.0.0
google-generativeai==0.3.2 